+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `Inline.BufferUtf8` property.
 - Added `InlineInterpolatedStringHandlerUtf8` struct.
 - Added overloads for formatting `System.Numerics.Plane` using interpolated string handlers.
+- Some types now implement `ISpanFormattable` and `IUtf8SpanFormattable`:
+  - `Rgba`
 
 ### Changed